Refactor `RobotModel` trait for unified modeling
The Robot trait does not need a Model type anymore. Instead, it uses the RobotModel struct, which is an abstraction over the model for panda and fr3.
Since currently fr3 and panda can work with the type, the RobotModel is also used internally. Should at a later point those two become incompatible, the internal model can be replaced with a type that converts into RobotModel